Move unit tests in separated file
