+++ src/ggrc-client/js/models/tests/cacheable_model_spec.js
@@ -90,53 +90,4 @@ describe('CMS.Models.Cacheable', function () {
       expect(instance._pending_joins.length).toEqual(1);
     });
   });
-
-  describe('cleanupACL() method', () => {
-    let resource;
-    let objectFromResourceSpy;
-    let model;
-    let id;
-
-    beforeEach(() => {
-      id = 711;
-      objectFromResourceSpy =
-        spyOn(can.Model.Cacheable, 'object_from_resource');
-      model = new can.Model.Cacheable({id: id});
-    });
-
-    afterEach(function () {
-      delete can.Model.Cacheable.cache;
-    });
-
-    it('returns resource if there is no object', () => {
-      resource = {};
-      objectFromResourceSpy.and.returnValue(undefined);
-
-      expect(can.Model.Cacheable.cleanupACL(resource)).toEqual(resource);
-    });
-
-    it('sets empty array to access_control_list attribute ' +
-    'if there is object with specified id in cache', () => {
-      model.attr('access_control_list', [1, 2, 3]);
-      resource = {id: id};
-      objectFromResourceSpy.and.returnValue(resource);
-
-      can.Model.Cacheable.cleanupACL(resource);
-      expect(model.attr('access_control_list').serialize())
-        .toEqual([]);
-    });
-
-    it('returns resource if there is no object with specified id in cache',
-      () => {
-        let acl = [1, 2, 3];
-        model.attr('access_control_list', acl);
-        resource = {id: id + 1};
-        objectFromResourceSpy.and.returnValue(resource);
-
-        can.Model.Cacheable.cleanupACL(resource);
-        expect(can.Model.Cacheable.cleanupACL(resource)).toEqual(resource);
-        expect(model.attr('access_control_list').serialize())
-          .toEqual(acl);
-      });
-  });
 });
